Renamed base package; added sub packages for app and domain
