+++ src/main/kotlin/uk/gov/justice/digital/hmpps/educationandworkplanapi/app/health/HealthInfo.kt
@@ -1,4 +1,4 @@
-package uk.gov.justice.digital.hmpps.hmppseducationandworkplanapi.health
+package uk.gov.justice.digital.hmpps.educationandworkplanapi.app.health
 
 import org.springframework.boot.actuate.health.Health
 import org.springframework.boot.actuate.health.HealthIndicator
