+++ src/main/kotlin/uk/gov/justice/digital/hmpps/educationandworkplanapi/app/health/VersionOutputter.kt
@@ -1,4 +1,4 @@
-package uk.gov.justice.digital.hmpps.hmppseducationandworkplanapi.health
+package uk.gov.justice.digital.hmpps.educationandworkplanapi.app.health
 
 import org.slf4j.LoggerFactory
 import org.springframework.boot.context.event.ApplicationReadyEvent
