+++ src/main/kotlin/uk/gov/justice/digital/hmpps/educationandworkplanapi/domain/Placeholder.kt
@@ -1,0 +1,4 @@
+package uk.gov.justice.digital.hmpps.educationandworkplanapi.domain
+
+// Placeholder file to maintain directory / package structure whilst there are no classes
+// Remove this file once we start adding classes
