+++ src/test/kotlin/uk/gov/justice/digital/hmpps/educationandworkplanapi/app/integration/IntegrationTestBase.kt
@@ -1,4 +1,4 @@
-package uk.gov.justice.digital.hmpps.hmppseducationandworkplanapi.integration
+package uk.gov.justice.digital.hmpps.educationandworkplanapi.app.integration
 
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.test.context.SpringBootTest
