+++ src/testFixtures/kotlin/uk/gov/justice/digital/hmpps/hmppseducationandworkplanapi/Placeholder.kt
@@ -1,0 +1,4 @@
+package uk.gov.justice.digital.hmpps.hmppseducationandworkplanapi
+
+// This file is only here as part of the initial commit in order to maintain the package directory structure
+// Remove once the first classes are added to this package
